split ws/http from `hub.rs`, hide `channel` from per-exchange
